+++ arenatacticsserver/scenes/actions/offensive_actions/melee_offensive_action.gd
@@ -1,6 +1,5 @@
-extends GameplayAction
-class_name BasicMovementAction
-
+extends BasicOffensiveAction
+class_name MeleeOffensiveAction
 
 func _init(instance: CharacterInstance, _data: Dictionary, _name: String):
 	super(instance, _data, _name)